Added Db context

+++ Backend/QuizPrototype.WebApi/QuizPrototype.Domain/Entities/BaseEntity.cs
@@ -3,7 +3,7 @@ namespace QuizPrototype.Domain.Entities
 {
     public class BaseEntity
     {
-        public virtual int Id { get; set; }
+        public virtual long Id { get; set; }
 
     }
 }
